update to v1.1
added ability to change some of the settings (were originally constants)
used in the animation. these allow the ability to change the strength
and length of the shake effect, the scale used for the letters, when the
fade starts and when to end the animation (length). also added ability
to specify a volume to allow control over the volume of the sound
without requiring manipulation of the code or the sound file. the volume
is specified as a multiplier (from zero to one).

the sound no longer stops automatically when the animation ends; it will
continue until it finishes or is stopped manually. this allows the
animation to finish before the sound while allowing the sound to finish
without cutting it off prematurely. the sound can still be stopped
manually (void stopSound()) if necessary. it is also possible to test to
see if the sound is still playing (bool isSoundFinished()).

updated example "main.cpp" to use the customisations. however, all of
the newly added customisations have default (the original) values if
none are given.

added changelog to repository.